# docs/club-scan-ordering/openapi.yaml
openapi: 3.0.3
info:
  title: PurelyClub 扫码点餐 API
  version: 0.1.0-draft
  description: |
    purelyClub 消费者端扫码点餐契约草案。
    所有金额原始值均为整数分；服务端是价格、优惠、库存、订单和支付状态的唯一权威来源。
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /club/scan-ordering/menu:
    get:
      tags: [Club Scan Ordering]
      summary: 获取当前会话菜单
      parameters:
        - $ref: '#/components/parameters/SessionIdQuery'
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 菜单数据
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
        '304':
          description: 菜单未变更
        '401':
          $ref: '#/components/responses/AuthRequired'
        '409':
          $ref: '#/components/responses/ScanSessionExpired'
  /club/scan-ordering/menu/products/{productId}:
    get:
      tags: [Club Scan Ordering]
      summary: 获取商品详情
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/SessionIdQuery'
      responses:
        '200':
          description: 商品详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuProduct'
        '404':
          $ref: '#/components/responses/ProductNotFound'
...
  /club/scan-ordering/service-calls:
    post:
      tags: [Club Scan Ordering]
      summary: 呼叫服务员
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceCallRequest'
      responses:
        '201':
          description: 呼叫已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCall'
        '429':
          $ref: '#/components/responses/ServiceCallTooFrequent'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionIdQuery:
      name: sessionId
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    AuthRequired:
      description: 需要登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            code: CLUB_AUTH_REQUIRED
            message: 请先登录后继续点餐。
            requestId: req_01J
            timestamp: '2026-07-23T12:00:00.000Z'
...
    ScanSessionExpired:
      description: 会话已过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ProductNotFound:
      description: 商品不存在或已下架
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceCallTooFrequent:
      description: 呼叫过于频繁
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    MenuResponse:
      type: object
      required: [menuVersion, categories]
      properties:
        menuVersion:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/MenuCategory'
    MenuCategory:
      type: object
      required: [id, name, sortOrder, products]
      properties:
        id:
          type: string
        name:
          type: string
        sortOrder:
          type: integer
        products:
          type: array
          items:
            $ref: '#/components/schemas/MenuProduct'
    MenuProduct:
      type: object
      required:
        [
          id,
          name,
          basePriceAmount,
          currency,
          stockStatus,
          isSoldOut,
          specGroups,
        ]
      properties:
        id:
          type: string
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        basePriceAmount:
          type: integer
          minimum: 0
        currency:
          type: string
          enum: [CNY]
        stockStatus:
          type: string
          enum: [available, low_stock, sold_out]
        isSoldOut:
          type: boolean
        specGroups:
          type: array
          items:
            $ref: '#/components/schemas/SpecGroup'
    SpecGroup:
      type: object
      required: [id, name, selectionType, minSelections, maxSelections, options]
      properties:
        id:
          type: string
        name:
          type: string
        selectionType:
          type: string
          enum: [single, multiple]
        minSelections:
          type: integer
          minimum: 0
        maxSelections:
          type: integer
          minimum: 1
        options:
          type: array
          items:
            $ref: '#/components/schemas/SpecOption'
    SpecOption:
      type: object
      required: [id, name, extraPriceAmount, isAvailable]
      properties:
        id:
          type: string
        name:
          type: string
        extraPriceAmount:
          type: integer
          minimum: 0
        isAvailable:
          type: boolean
...
    CreateServiceCallRequest:
      type: object
      required: [sessionId, type]
      properties:
        sessionId:
          type: string
        type:
          type: string
          enum: [waiter, water, cutlery, checkout, other]
        remark:
          type: string
          maxLength: 200
    ServiceCall:
      type: object
      required: [id, type, status, requestedAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [waiter, water, cutlery, checkout, other]
        status:
          type: string
          enum: [pending, acknowledged, resolved, cancelled]
        requestedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [statusCode, code, message, requestId, timestamp]
      properties:
        statusCode:
          type: integer
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time